// src/app/core/models/profile.ts
export class Profile {
    id: number;
    title: string;
    firstname: string;
    lastname: string;
    birthday: Date;
    address: string;
    availability: string;
    scope: string;
    license: string;
    jobTitle: string;
    facebook: string;
    linkedin: string;
    email: string;
    active = false;
}